# api.yaml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/SMARTBEARD192M/Casita/1.0.0
  - description: Development server
    url: http://localhost:8001
info:
  description: API for my Casita household management app
  version: "1.0.0"
  title: Casita API
  contact:
    email: andresap@protonmail.com
    name: Andrés Arango Pérez
  license:
    name: MIT
    url: https://www.mit.edu/~amini/LICENSE.md
tags:
  - name: households
    description: Household related calls
paths:
  /households:
    get:
      operationId: getUserHouseholds
      tags:
        - households
      summary: Get all households associated to a user
      parameters:
        - name: userId
          in: query
          description: Id of the household owner, whose households will be returned
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Household'
        '500':
          description: Internal server error
        '400':
          description: Invalid owner ID
    post:
      operationId: createHousehold
      tags:
        - households
      summary: Create a new household
      requestBody:
        description: Household data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Household'
      responses:
        '204':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsertedId'
        '400':
          description: Validation error or bad request
        '500':
          description: Internal server error
components:
  schemas:
    Household:
      type: object
      properties:
        id:
          type: string
        ownerId:
          type: integer
          format: int64
        street:
          type: string
        houseNumber:
          type: string
        postalCode:
          type: string
        city:
          type: string
        country:
          type: string
      required:
        - ownerId
        - street
        - houseNumber
    InsertedId:
      type: object
      properties:
        insertedId:
          type: string
          description: ID of a newly inserted document to MongoDB